% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dHN.R
\name{dHN}
\alias{dHN}
\alias{dHN_V}
\alias{rHN}
\alias{rHN_V}
\title{Half-normal distribution for use in \code{nimble} models}
\usage{
dHN(x, sigma, Xmax = 100, log = 0)

rHN(n, sigma, Xmax = 100)

dHN_V(x, sigma, Xmax = 100, log = 0)

rHN_V(n, sigma, Xmax = 100)
}
\arguments{
\item{x}{distance data. either a single value (dHN) or a vector of values (dHN_V)}

\item{sigma}{scale of the half-normal distribution}

\item{Xmax}{right truncation distance for integration of the likelihood function}

\item{log}{if TRUE, return the log-likelihood}

\item{Xmin}{left truncation distance for integration of the likelihood function}
}
\description{
\code{dHN} and \code{dHN_V} provide hazard-rate detection
distributions that can be used directly from R or in \code{nimble}
models. The corresponding randomisation routines \code{rHN} and \code{rHN_V} generate random distances
from the corresponding distance-detection distribution.
}
\examples{

#direct invocation of functions from R to evaluate the likelihood
dHN(x=20,  sigma=40, Xmax=100)
dHN(x=20,  sigma=40, Xmax=100)

N=500
true_y<-runif(N, 0, 100)
#hazard rate truth
sigma_true<-40
#half-normal detection
p_detect<- exp(-true_y^2/sigma_true^2)
plot(p_detect~true_y, type="p")
detect<-rbinom(N, 1, p_detect)
#observations
y<-true_y[detect==1]
nind<-length(y)
hist(y, breaks=20)
distCodeV<-nimbleCode({
 y[]~dHN_V(sigma, 100)
 sigma~dunif(10, 70)
 })
#inits and monitors
inits <- function() list(sigma=50)
params <- c("sigma")

#generate some MCMC samples
samples <- nimbleMCMC(
code = distCodeV,
constants = list(y=y, nind=nind), ## the distance data
inits = inits,
monitors = params,
niter = 1000,
nburnin = 500)

layout(matrix(1:3, nrow=3))
hist(samples[,"sigma"], col="red", xlab="sigma")
abline(v=sigma_true, col="blue", lwd=2)
hist(samples[,"p_mean"], col="red", xlab="p_mean", xlim=c(0, 1))
abline(v=RintegralHR(b_true, sigma_true, 0, 100)/100, col="blue", lwd=2)
}
\author{
Michael Scroggie
}
\keyword{hazard}
\keyword{rate}
