% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dHN.R
\name{dHN}
\alias{dHN}
\alias{integralHN}
\alias{dHN_V}
\alias{rHN}
\alias{rHN_V}
\title{Half-normal distribution for use in \code{nimble} models}
\usage{
integralHN(sigma = 30, Xmin, Xmax, point = 0)

dHN(x, sigma, Xmax = 100, point = 0, log = 0)

rHN(n, sigma, Xmax = 100, point = 0)

dHN_V(x, sigma, Xmax = 100, point = 0, log = 0)

rHN_V(n, sigma, Xmax = 100, point = 0)
}
\arguments{
\item{sigma}{scale of the half-normal distribution}

\item{Xmin}{left truncation distance for integration of the likelihood function}

\item{Xmax}{right truncation distance for integration of the likelihood function}

\item{point}{logical, if 1 compute likelihood for point transects, if 0 (default) compute likelihood for line transects}

\item{x}{distance data. either a single value (dHN) or a vector of values (dHN_V)}

\item{log}{if TRUE, return the log-likelihood}
}
\description{
\code{dHN} and \code{dHN_V} provide hazard-rate detection
distributions that can be used directly from R or in \code{nimble}
models. The corresponding randomisation routines \code{rHN} and \code{rHN_V} generate random distances
from the corresponding distance-detection distribution.
\code{integralHN} is a helper function for computing the integral for the half-normal function with the specified parameters.
}
\examples{

#direct invocation of functions from R to evaluate the likelihood
dHN(x=20,  sigma=40, Xmax=100)
dHN(x=20,  sigma=40, Xmax=100, point=1)

N=1000
true_y<-runif(N, 0, 100)
#hazard rate truth
sigma_true<-40
#half-normal detection
p_detect<- exp(-true_y^2/sigma_true^2)
plot(p_detect~true_y, type="p")
detect<-rbinom(N, 1, p_detect)
#observations
y<-true_y[detect==1]
nind<-length(y)
hist(y, breaks=20)
distCodeV<-nimbleCode({
 y[1:nind]~dHN_V(sigma, 100, point=0)
 sigma~dunif(10, 70)
 })
#inits and monitors
inits <- function() list(sigma=50)
params <- c("sigma")

#generate some MCMC samples
samples <- nimbleMCMC(
code = distCodeV,
constants = list(y=y, nind=nind), ## the distance data
inits = inits,
monitors = params,
niter = 2000,
nburnin = 1000)

layout(matrix(1:2, nrow=2))
hist(y, freq=FALSE)
lines(x=1:100, y=dHN(1:100, mean(samples[,"sigma"])), col="red")
hist(samples[,"sigma"], col="red")
}
\author{
Michael Scroggie
}
\keyword{hazard}
\keyword{rate}
