% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dHR.R
\name{dHR}
\alias{dHR}
\alias{integralHR}
\alias{dHR_V}
\alias{rHR}
\alias{rHR_V}
\title{Hazard rate distribution for use in \code{nimble} models}
\usage{
integralHR(b = 5, sigma = 30, Xmin, Xmax)

dHR(x, b, sigma, Xmax = 100, log = 0)

rHR(n, b, sigma, Xmax = 100)
}
\arguments{
\item{b}{shape of the hazard rate function}

\item{sigma}{scale of the hazard rate function}

\item{Xmax}{truncation distance for integration of the likelihood function}

\item{x}{distance data. either a single value (dHR) or a vector of values (dHR_V)}

\item{log}{if TRUE, return the log-likelihood}
}
\description{
\code{dHR} and \code{dHR_V} provide hazard-rate detection
distributions that can be used directly from R or in \code{nimble}
models.
}
\author{
Michael Scroggie
}
\keyword{hazard}
\keyword{rate}
